increase OTM coverage to diagrams
