+++ .gitignore
@@ -81,7 +81,9 @@
 !td.vue/src/service/demo/
 !td.vue/src/service/entity/
 !td.vue/src/service/migration/
+!td.vue/src/service/schema/
 !td.vue/src/service/provider/
+!td.vue/src/service/otm/
 !td.vue/src/service/threats/
 !td.vue/src/service/threats/models/
 !td.vue/src/service/x6/
@@ -113,9 +115,11 @@
 !td.vue/tests/unit/service/demo/
 !td.vue/tests/unit/service/entity/
 !td.vue/tests/unit/service/migration/
+!td.vue/tests/unit/service/otm/
 !td.vue/tests/unit/service/provider/
 !td.vue/tests/unit/service/threats/
 !td.vue/tests/unit/service/threats/models/
+!td.vue/tests/unit/service/schema/
 !td.vue/tests/unit/service/x6/
 !td.vue/tests/unit/service/x6/graph/
 !td.vue/tests/unit/service/x6/shapes/
