+++ td.vue/src/service/otm/TDtoOTM.js
@@ -1,5 +1,3 @@
-const buildVersion = require('../../../package.json').version;
-
 const convertTDtoOTM = (data) => {
     const jsonData = JSON.stringify(data, null, 2);
 
@@ -223,174 +221,6 @@ const convertTDtoOTM = (data) => {
     return otm;
 };
 
-const convertOTMtoTD = (jsonModel) => {
-    console.log(jsonModel);
-    const noteText = 'Note that Open Threat Model is not supported, yet.\n ';
-
-    var dragonModel = new Object();
-    dragonModel.version = buildVersion;
-    dragonModel.otmVersion = jsonModel.otmVersion;
-
-    dragonModel.summary = new Object();
-    dragonModel.summary.title = jsonModel.project.name;
-    dragonModel.summary.owner = jsonModel.project.owner;
-    dragonModel.summary.ownerContact = jsonModel.project.ownerContact;
-    dragonModel.summary.description = noteText + jsonModel.project.description;
-    dragonModel.summary.id = jsonModel.project.id;
-
-    // both attributes and tags are not used yet by TD, but need to be preserved if present
-    if (jsonModel.project.attributes) {
-        dragonModel.attributes = JSON.parse(JSON.stringify(jsonModel.project.attributes));
-    }
-    if (jsonModel.project.tags) {
-        dragonModel.tags = JSON.parse(JSON.stringify(jsonModel.project.tags));
-    }
-
-    dragonModel.detail = new Object();
-    dragonModel.detail.contributors = [];
-    dragonModel.detail.diagrams = [];
-    dragonModel.detail.diagramTop = 0;
-    dragonModel.detail.reviewer = '';
-    dragonModel.detail.threatTop = 0;
-    console.log(dragonModel);
-
-    /*
-    var diagram = new Object();
-    diagram.version = '2.0';
-    diagram.title = 'Main Request Data Flow';
-    diagram.id = 0;
-    diagram.diagramType = 'STRIDE';
-    diagram.cells = [];
-
-    jsonModel.components.forEach(function(comp) {
-        var cell = new Object();
-        cell.shape = 'process';
-        cell.zIndex = 1;
-        cell.id = comp.id;
-        cell.data = new Object();
-        cell.data.name = comp.name;
-        cell.data.type = comp.representations[0].name;
-        cell.position = new Object();
-        cell.position.x = comp.representations[0].position.x;
-        cell.position.y = comp.representations[0].position.y;
-        cell.size = new Object();
-        cell.size.width = comp.representations[0].size.width;
-        cell.size.height = comp.representations[0].size.height;
-        cell.attrs = new Object();
-        cell.attrs.text = new Object();
-        cell.attrs.text.text = comp.name;
-
-        if (Object.hasOwn(comp, 'threats')) {
-            cell.data.threats = [];
-            comp.threats.forEach(function(threat) {
-                var t = new Object();
-                t.title = threat.name;
-                t.id = threat.id;
-                t.description = threat.description;
-                t.status = threat.attributes[0].status;
-                t.severity = threat.attributes[0].severity;
-                t.mitigation = threat.attributes[0].mitigation;
-                t.type = threat.attributes[0].type;
-                t.modelType = threat.attributes[0].modelType;
-                cell.data.threats.push(t);
-            });
-        }
-
-        diagram.cells.push(cell);
-    });
-
-    jsonModel.dataflows.forEach(function(flow) {
-        var cell = new Object();
-        cell.shape = 'flow';
-        cell.zIndex = 10;
-        cell.id = flow.id;
-        cell.data = new Object();
-        cell.data.name = flow.name;
-        cell.connector = 'smooth';
-        if (Object.hasOwn(flow, 'representations')) {
-            cell.data.type = flow.representations[0].name;
-        }
-        cell.width = 200;
-        cell.height = 100;
-        cell.source = new Object();
-        cell.source.cell = flow.source;
-        cell.target = new Object();
-        cell.target.cell = flow.destination;
-        cell.vertices = [];
-        if (Object.hasOwn(flow, 'attributes')) {
-            var vert = new Object();
-            vert.x = flow.attributes[0].verticesX;
-            vert.y = flow.attributes[0].verticesY;
-            cell.vertices.push(vert);
-        }
-        cell.labels = [];
-        var lbl = new Object();
-        lbl.attrs = new Object();
-        lbl.attrs.label = new Object();
-        lbl.attrs.label.text = flow.name;
-        cell.labels.push(lbl);
-
-        if (Object.hasOwn(flow, 'threats')) {
-            cell.data.threats = [];
-            flow.threats.forEach(function(threat) {
-                var t = new Object();
-                t.title = threat.name;
-                t.id = threat.id;
-                t.description = threat.description;
-                t.status = threat.attributes[0].status;
-                t.severity = threat.attributes[0].severity;
-                t.mitigation = threat.attributes[0].mitigation;
-                t.type = threat.attributes[0].type;
-                t.modelType = threat.attributes[0].modelType;
-                cell.data.threats.push(t);
-            });
-        }
-
-        diagram.cells.push(cell);
-    });
-
-    jsonModel.trustZones.forEach(function(zone) {
-        var cell = new Object();
-        cell.shape = 'trust-broundary-curve';
-        cell.zIndex = 10;
-        cell.id = zone.id;
-        cell.data = new Object();
-        cell.data.name = zone.name;
-        cell.connector = 'smooth';
-        cell.width = 200;
-        cell.height = 100;
-        if (Object.hasOwn(zone, 'attributes')) {
-            cell.data.type = zone.attributes[0].type;
-            cell.source = new Object();
-            cell.source.x = zone.attributes[0].sourceX;
-            cell.source.y = zone.attributes[0].sourceY;
-            cell.target = new Object();
-            cell.target.x = zone.attributes[0].targetX;
-            cell.target.y = zone.attributes[0].targetY;
-        }
-        cell.vertices = [];
-        zone.representations.forEach(function(rep) {
-            var vert = new Object();
-            vert.x = rep.position.x;
-            vert.y = rep.position.y;
-            cell.vertices.push(vert);
-        });
-        cell.labels = [];
-        var lbl = new Object();
-        lbl.attrs = new Object();
-        lbl.attrs.label = new Object();
-        lbl.attrs.label.text = zone.name;
-        cell.labels.push(lbl);
-        diagram.cells.push(cell);
-    });
-
-    dragonModel.detail.diagrams.push(diagram);
-
-*/
-    return dragonModel;
-};
-
 export default {
-    convertOTMtoTD,
     convertTDtoOTM
 };
