+++ td.vue/src/service/otm/openThreatModel.js
@@ -1,0 +1,7 @@
+import convertTDtoOTM from './TDtoOTM.js';
+import convertOTMtoTD from './OTMtoTD.js';
+
+export default {
+    convertOTMtoTD,
+    convertTDtoOTM
+};
