+++ td.vue/src/service/save.js
@@ -1,35 +1,14 @@
-import openThreatModel from '../service/openThreatModel.js';
-
 /**
  * Saves the file locally
  * @param {Object} data
  * @param {string} fileName
- * @param {string} format
  */
-const local = (data, fileName, format) => {
-    if (format == 'otm'){
-        saveOTM(data, fileName);
-    }
-    else{
-        saveTD(data, fileName);
-    }
+const local = (data, fileName) => {
+    saveTD(data, fileName);
 };
 
-function saveOTM (data, fileName) {
-    var otm = openThreatModel.convertTDtoOTM(data);
-    var saveOTM = JSON.stringify(otm);
-
-    const contentType = 'application/json';
-    const otmblob = new Blob([saveOTM], { type: contentType });
-    const a = document.createElement('a');
-    a.href = window.URL.createObjectURL(otmblob);
-    a.download = fileName;
-    a.click();
-    a.remove();
-}
-
 function saveTD (data, fileName) {
-    // TODO: Split into local browser and electron-specific saving
+    // saving using local browser  
     const contentType = 'application/json';
     const jsonData = JSON.stringify(data, null, 2);
     const blob = new Blob([jsonData], { type: contentType });
